@chris-morgan 2015-02-03
(I believe a certain bluntness is needed to be effective in this case. Pray take no offence, any one who should read this.)

Speaking of the *status quo* with no reference to the changes enacted in this PR:

I’ve never looked closely at what hyper did with its header representation and mucell and so forth. I took a little closer look just now and I am not at all pleased at what I see.

Header reparsing (accessing it as multiple types) is not permitted, which while being more restrictive in corner cases does simplify various things. Most notably, it does indeed mean that mutation can be constrained to only occur through a mutable reference, and so indeed mucell’s niftiness is wasted. I deemed the misbehaviour in corner cases as unacceptable in Teepee, which is why I wrote mucell—but it does have additional costs, most notably that you cannot use plain references any more but must propagate new reference types that implement `Deref`. If you consider the corner-case misbehaviour acceptable, then you might as well bypass the small costs of mucell.

To the author or authors of all this code: your license to use the `unsafe` keyword and more importantly the `transmute` function has been revoked for 30 days and you will be on probation for 180 days thereafter. Injudicious usage of transmutation has left the header representation subsystem with notable memory safety bugs. This is in no manner the fault of mucell but rather the cruel and unusual treatment it has received (e.g. `mem::transmute::<&H, &H>(downcast(&*item.borrow()))`, ruining the reference counting, about which I may furthermore add that the `downcast(&*item.borrow())` part should not compile due to a lifetime conflict). As mentioned above, if you’re going to use mucell returning plain references is not an option; new reference types to handle the dropping of the reference count must be used, as is demonstrated in Teepee, the header representation scheme of which I believe to be devoid of memory-safety bugs.

I have not extensively assessed the impact of this PR, but I do know that my judgement on the issue is that the entire header module needs careful auditing and probably rewriting and I believe all unsafe code should be removed from the module with prejudice for the moment.

I will leave you with just one comment about the PR itself:

Bear in mind also that this `OptCell` is *not* a safe type; a reference may exist through `Deref` when `set` is called. `set` should therefore be marked an unsafe function.