@chris-morgan 2015-02-03
I mostly retract my last paragraph; because `set` can only be called when the wrapped value is `None` (in debug mode, anyway) the worst that can happen is that a supposedly immutable `&None` changes underneath the variable binding to `&Some(…)`. This is bad, but does not affect memory safety as there can be no references to inside a `None`.