@seanmonstar 2015-02-03
> Header reparsing (accessing it as multiple types) is not permitted, which while being more restrictive in corner cases does simplify various things.

I took the road of "we don't need it yet, and it grossly complicates things. Let's not have it for now, and if that's a terrible thing, people will file issues."

> To the author or authors of all this code: your license to use the unsafe keyword and more importantly the transmute function has been revoked for 30 days and you will be on probation for 180 days thereafter.

This was the theatrics I mentioned. The rest of the critique is perfectly fine.

> Injudicious usage of transmutation has left the header representation subsystem with notable memory safety bugs.

The unsafety is not as dire as it seems. Parsing can only happen once, and so the &H received cannot be changed from under you. The only way to change the value is by using a &mut method, which the borrow checker secures for us.

If an actual example of the unsafety can be shown, I'd love to learn from it.

> I have not extensively assessed the impact of this PR, but I do know that my judgement on the issue is that the entire header module needs careful auditing and probably rewriting and I believe all unsafe code should be removed from the module with prejudice for the moment.

At this point, the unsafe code that exists at this point is: a) downcast_unchecked, b) OptCell, exposing a safe exterior, and c) one lifetime transmute to handle an annoyance with `HashMap<CowString<'static>, V>`.

> Bear in mind also that this OptCell is not a safe type; a reference may exist through Deref when set is called. set should therefore be marked an unsafe function.

`set()` can only be called once. An assertion exists that the `Option.is_none()` before setting. You could take a reference to the `None`, but we don't. The only references returned are if the value is `Some`.

Any mutable access to the reference requires both an `&mut` method, and resetting the opposite value (raw or typed) so they stay in sync.