@seanmonstar 2015-02-07
Merging since this doesn't change the semantics of the design, makes things faster and safer, and doesn't prevent #293. r=Manish